doors are a thing now.
closed door is a nonwall object that can be bumped into to open.
open door is a nonwall object that can be walked into.

a closed door becomes an open door when it is bumped into.

fixed bug where if you traveled in a straight line after coming through a board door and you bumped into a colidable nonwall collision would get buggy.